Add run command

+++ src/esperoj/esperoj.py
@@ -54,7 +54,14 @@ class Esperoj:
             "Authorization": f"LOW {api_key}:{api_secret}",
         }
 
-        params = {"url": url, "skip_first_archive": 1, "force_get": 1, "email_result": 1}
+        params = {
+            "url": url,
+            "skip_first_archive": 1,
+            "force_get": 1,
+            "email_result": 1,
+            "delay_wb_availability": 0,
+            "capture_screenshot": 0,
+        }
         response = requests.post("https://web.archive.org/save", headers=headers, data=params)
         if response.status_code != 200:
             raise RuntimeError(f"Error: {response.text}")
@@ -107,13 +114,12 @@ class Esperoj:
         Raises:
             FileNotFoundError: If the file is not in the repository.
         """
-        files = self.db.table("Files")
-        record = next(files.get_all({"Name": name}), None)
+        record = next(self.db.table("Files").get_all({"Name": name}), None)
         if not record:
             raise FileNotFoundError
         url = self.storage.get_link(name)
         archive_url = Esperoj._archive(url)
-        return files.update(record.record_id, {"Internet Archive": archive_url})
+        return record.update({"Internet Archive": archive_url})
 
     def ingest(self, path: Path) -> Record:
         """Ingest the file at the given path into the database and the storage.
@@ -131,16 +137,15 @@ class Esperoj:
         if not path.is_file():
             raise FileNotFoundError
 
-        files = self.db.table("Files")
         name = path.name
-
-        if list(files.get_all({"Name": name})) != []:
-            raise FileExistsError
-
-        if self.storage.file_exists(name):
-            raise FileExistsError
         size = path.stat().st_size
         sha256sum = Esperoj._calculate_hash(path.read_bytes(), algorithm="sha256")
+        files = self.db.table("Files")
+
+        if self.storage.file_exists(name) or (
+            next(files.get_all({"Name": name}), None) is not None
+        ):
+            raise FileExistsError
 
         self.storage.upload_file(str(path), name)
